quality: Fix many classes of warnings
First step in getting -Werror to work after the refactor.

+++ src/peerstimelinemodel.cpp
@@ -173,7 +173,7 @@ int PeersTimelineModel::columnCount( const QModelIndex& parent) const
 
 QModelIndex PeersTimelineModel::index(int row, int col, const QModelIndex& p) const
 {
-    return (row < 0 || row >= d_ptr->m_lRows.size() || p.isValid() || col > 3) ?
+    return (row < 0 || row >= (int)d_ptr->m_lRows.size() || p.isValid() || col > 3) ?
         QModelIndex() : createIndex(row, col, d_ptr->m_lRows[d_ptr->realIndex(row)]);
 }
